+++ metricflow/dataflow/dataflow_plan.py
@@ -3,42 +3,34 @@
 from __future__ import annotations
 
 import logging
-import textwrap
+import typing
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from typing import Generic, List, Optional, Sequence, Tuple, Type, TypeVar, Union
+from typing import Generic, Optional, Sequence, Type, TypeVar
 
-import jinja2
-from dbt_semantic_interfaces.protocols.metric import MetricTimeWindow
-from dbt_semantic_interfaces.references import TimeDimensionReference
-from dbt_semantic_interfaces.type_enums.aggregation_type import AggregationType
-from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
-
-from metricflow.dag.id_prefix import IdPrefix, StaticIdPrefix
-from metricflow.dag.mf_dag import DagId, DagNode, DisplayedProperty, MetricFlowDag, NodeId
-from metricflow.dataflow.builder.partitions import (
-    PartitionDimensionJoinDescription,
-    PartitionTimeDimensionJoinDescription,
-)
-from metricflow.dataflow.sql_table import SqlTable
-from metricflow.dataset.sql_dataset import SqlDataSet
-from metricflow.filters.time_constraint import TimeRangeConstraint
-from metricflow.mf_logging.pretty_print import mf_pformat
-from metricflow.specs.specs import (
-    ConstantPropertySpec,
-    EntitySpec,
-    InstanceSpec,
-    InstanceSpecSet,
-    LinklessEntitySpec,
-    MeasureSpec,
-    MetricInputMeasureSpec,
-    MetricSpec,
-    OrderBySpec,
-    TimeDimensionSpec,
-    WhereFilterSpec,
-)
-from metricflow.sql.sql_plan import SqlJoinType
+from metricflow.dag.id_prefix import StaticIdPrefix
+from metricflow.dag.mf_dag import DagId, DagNode, MetricFlowDag, NodeId
 from metricflow.visitor import Visitable, VisitorOutputT
+
+if typing.TYPE_CHECKING:
+    from metricflow.dataflow.nodes.add_generated_uuid import AddGeneratedUuidColumnNode
+    from metricflow.dataflow.nodes.aggregate_measures import AggregateMeasuresNode
+    from metricflow.dataflow.nodes.combine_aggregated_outputs import CombineAggregatedOutputsNode
+    from metricflow.dataflow.nodes.compute_metrics import ComputeMetricsNode
+    from metricflow.dataflow.nodes.constrain_time import ConstrainTimeRangeNode
+    from metricflow.dataflow.nodes.filter_elements import FilterElementsNode
+    from metricflow.dataflow.nodes.join_conversion_events import JoinConversionEventsNode
+    from metricflow.dataflow.nodes.join_over_time import JoinOverTimeRangeNode
+    from metricflow.dataflow.nodes.join_to_base import JoinToBaseOutputNode
+    from metricflow.dataflow.nodes.join_to_time_spine import JoinToTimeSpineNode
+    from metricflow.dataflow.nodes.metric_time_transform import MetricTimeDimensionTransformNode
+    from metricflow.dataflow.nodes.min_max import MinMaxNode
+    from metricflow.dataflow.nodes.order_by_limit import OrderByLimitNode
+    from metricflow.dataflow.nodes.read_sql_source import ReadSqlSourceNode
+    from metricflow.dataflow.nodes.semi_additive_join import SemiAdditiveJoinNode
+    from metricflow.dataflow.nodes.where_filter import WhereConstraintNode
+    from metricflow.dataflow.nodes.write_to_dataframe import WriteToResultDataframeNode
+    from metricflow.dataflow.nodes.write_to_table import WriteToResultTableNode
+
 
 logger = logging.getLogger(__name__)
 
@@ -186,254 +178,6 @@ class BaseOutput(DataflowPlanNode, ABC):
     pass
 
 
-class ReadSqlSourceNode(BaseOutput):
-    """A source node where data from an SQL table or SQL query is read and output."""
-
-    def __init__(self, data_set: SqlDataSet) -> None:
-        """Constructor.
-
-        Args:
-            data_set: dataset describing the SQL table / SQL query
-        """
-        self._dataset = data_set
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=())
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_READ_SQL_SOURCE_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_source_node(self)
-
-    @property
-    def data_set(self) -> SqlDataSet:
-        """Return the data set that this source represents and is passed to the child nodes."""
-        return self._dataset
-
-    def __str__(self) -> str:  # noqa: D
-        return jinja2.Template(
-            textwrap.dedent(
-                """\
-                <{{ class_name }} data_set={{ data_set }} />
-                """
-            )
-        ).render(class_name=self.__class__.__name__, data_set=str(self.data_set))
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return f"""Read From {self.data_set}"""
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return tuple(super().displayed_properties) + (DisplayedProperty("data_set", self.data_set),)
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return isinstance(other_node, self.__class__) and other_node.data_set == self.data_set
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> ReadSqlSourceNode:  # noqa: D
-        assert len(new_parent_nodes) == 0
-        return ReadSqlSourceNode(data_set=self.data_set)
-
-
-@dataclass(frozen=True)
-class ValidityWindowJoinDescription:
-    """Encapsulates details about join constraints around validity windows."""
-
-    window_start_dimension: TimeDimensionSpec
-    window_end_dimension: TimeDimensionSpec
-
-
-@dataclass(frozen=True)
-class JoinDescription:
-    """Describes how data from a node should be joined to data from another node."""
-
-    join_node: BaseOutput
-    join_on_entity: Optional[LinklessEntitySpec]
-    join_type: SqlJoinType
-
-    join_on_partition_dimensions: Tuple[PartitionDimensionJoinDescription, ...]
-    join_on_partition_time_dimensions: Tuple[PartitionTimeDimensionJoinDescription, ...]
-
-    validity_window: Optional[ValidityWindowJoinDescription] = None
-
-    def __post_init__(self) -> None:  # noqa: D
-        if self.join_on_entity is None and self.join_type != SqlJoinType.CROSS_JOIN:
-            raise RuntimeError("`join_on_entity` is required unless using CROSS JOIN.")
-
-
-class JoinToBaseOutputNode(BaseOutput):
-    """A node that joins data from other nodes to a standard output node, one by one via entity."""
-
-    def __init__(
-        self,
-        left_node: BaseOutput,
-        join_targets: Sequence[JoinDescription],
-        node_id: Optional[NodeId] = None,
-    ) -> None:
-        """Constructor.
-
-        Args:
-            left_node: node with standard output
-            join_targets: other sources that should be joined to this node.
-            node_id: Override the node ID with this value
-        """
-        self._left_node = left_node
-        self._join_targets = tuple(join_targets)
-
-        # Doing a list comprehension throws a type error, so doing it this way.
-        parent_nodes: List[DataflowPlanNode] = [self._left_node]
-        for join_target in self._join_targets:
-            parent_nodes.append(join_target.join_node)
-        super().__init__(node_id=node_id or self.create_unique_id(), parent_nodes=parent_nodes)
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_JOIN_TO_STANDARD_OUTPUT_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_join_to_base_output_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return """Join Standard Outputs"""
-
-    @property
-    def left_node(self) -> BaseOutput:  # noqa: D
-        return self._left_node
-
-    @property
-    def join_targets(self) -> Sequence[JoinDescription]:  # noqa: D
-        return self._join_targets
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return tuple(super().displayed_properties) + tuple(
-            DisplayedProperty(f"join{i}_for_node_id_{join_description.join_node.node_id}", join_description)
-            for i, join_description in enumerate(self._join_targets)
-        )
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        if not isinstance(other_node, self.__class__) or len(self.join_targets) != len(other_node.join_targets):
-            return False
-
-        for i in range(len(self.join_targets)):
-            if (
-                self.join_targets[i].join_on_entity != other_node.join_targets[i].join_on_entity
-                or self.join_targets[i].join_on_partition_dimensions
-                != other_node.join_targets[i].join_on_partition_dimensions
-                or self.join_targets[i].join_on_partition_time_dimensions
-                != other_node.join_targets[i].join_on_partition_time_dimensions
-                or self.join_targets[i].validity_window != other_node.join_targets[i].validity_window
-            ):
-                return False
-        return True
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> JoinToBaseOutputNode:  # noqa: D
-        assert len(new_parent_nodes) > 1
-        new_left_node = new_parent_nodes[0]
-        new_join_nodes = new_parent_nodes[1:]
-        assert len(new_join_nodes) == len(self._join_targets)
-
-        return JoinToBaseOutputNode(
-            left_node=new_left_node,
-            join_targets=[
-                JoinDescription(
-                    join_node=new_join_nodes[i],
-                    join_on_entity=old_join_target.join_on_entity,
-                    join_on_partition_dimensions=old_join_target.join_on_partition_dimensions,
-                    join_on_partition_time_dimensions=old_join_target.join_on_partition_time_dimensions,
-                    validity_window=old_join_target.validity_window,
-                    join_type=old_join_target.join_type,
-                )
-                for i, old_join_target in enumerate(self._join_targets)
-            ],
-        )
-
-
-class JoinOverTimeRangeNode(BaseOutput):
-    """A node that allows for cumulative metric computation by doing a self join across a cumulative date range."""
-
-    def __init__(
-        self,
-        parent_node: BaseOutput,
-        time_dimension_spec_for_join: TimeDimensionSpec,
-        window: Optional[MetricTimeWindow],
-        grain_to_date: Optional[TimeGranularity],
-        node_id: Optional[NodeId] = None,
-        time_range_constraint: Optional[TimeRangeConstraint] = None,
-    ) -> None:
-        """Constructor.
-
-        Args:
-            parent_node: node with standard output
-            window: time window to join over
-            grain_to_date: indicates time range should start from the beginning of this time granularity
-            (eg month to day)
-            node_id: Override the node ID with this value
-            time_range_constraint: time range to aggregate over
-            time_dimension_spec_for_join: time dimension spec to use when joining to time spine
-        """
-        if window and grain_to_date:
-            raise RuntimeError(
-                f"This node cannot be initialized with both window and grain_to_date set. This configuration should "
-                f"have been prevented by model validation. window: {window}. grain_to_date: {grain_to_date}."
-            )
-        self._parent_node = parent_node
-        self._grain_to_date = grain_to_date
-        self._window = window
-        self.time_range_constraint = time_range_constraint
-        self.time_dimension_spec_for_join = time_dimension_spec_for_join
-
-        # Doing a list comprehension throws a type error, so doing it this way.
-        parent_nodes: Sequence[DataflowPlanNode] = [self._parent_node]
-        super().__init__(node_id=node_id or self.create_unique_id(), parent_nodes=parent_nodes)
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_JOIN_SELF_OVER_TIME_RANGE_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_join_over_time_range_node(self)
-
-    @property
-    def grain_to_date(self) -> Optional[TimeGranularity]:  # noqa: D
-        return self._grain_to_date
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return """Join Self Over Time Range"""
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        return self._parent_node
-
-    @property
-    def window(self) -> Optional[MetricTimeWindow]:  # noqa: D
-        return self._window
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return super().displayed_properties
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return (
-            isinstance(other_node, self.__class__)
-            and other_node.grain_to_date == self.grain_to_date
-            and other_node.window == self.window
-            and other_node.time_range_constraint == self.time_range_constraint
-        )
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> JoinOverTimeRangeNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return JoinOverTimeRangeNode(
-            parent_node=new_parent_nodes[0],
-            window=self.window,
-            grain_to_date=self.grain_to_date,
-            time_range_constraint=self.time_range_constraint,
-            time_dimension_spec_for_join=self.time_dimension_spec_for_join,
-        )
-
-
 class AggregatedMeasuresOutput(BaseOutput, ABC):
     """A node that outputs data where the measures are aggregated.
 
@@ -443,519 +187,10 @@ class AggregatedMeasuresOutput(BaseOutput, ABC):
     pass
 
 
-class AggregateMeasuresNode(AggregatedMeasuresOutput):
-    """A node that aggregates the measures by the associated group by elements.
-
-    In the event that one or more of the aggregated input measures has an alias assigned to it, any output query
-    resulting from an operation on this node must apply the alias and transform the measure instances accordingly,
-    otherwise this join could produce a query with two identically named measure columns with, e.g., different
-    constraints applied to the measure.
-    """
-
-    def __init__(
-        self,
-        parent_node: BaseOutput,
-        metric_input_measure_specs: Sequence[MetricInputMeasureSpec],
-    ) -> None:
-        """Initializer for AggregateMeasuresNode.
-
-        The input measure specs are required for downstream nodes to be aware of any input measures with
-        user-provided aliases, such as we might encounter with constrained and unconstrained versions of the
-        same input measure.
-        """
-        self._parent_node = parent_node
-        self._metric_input_measure_specs = tuple(metric_input_measure_specs)
-
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(self._parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_AGGREGATE_MEASURES_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_aggregate_measures_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return """Aggregate Measures"""
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        return self._parent_node
-
-    @property
-    def metric_input_measure_specs(self) -> Tuple[MetricInputMeasureSpec, ...]:
-        """Iterable of specs for measure inputs to downstream metrics.
-
-        Used for assigning aliases to output columns produced by aggregated measures.
-        """
-        return self._metric_input_measure_specs
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return (
-            isinstance(other_node, self.__class__)
-            and other_node.metric_input_measure_specs == self.metric_input_measure_specs
-        )
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> AggregateMeasuresNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return AggregateMeasuresNode(
-            parent_node=new_parent_nodes[0],
-            metric_input_measure_specs=self.metric_input_measure_specs,
-        )
-
-
-class SemiAdditiveJoinNode(BaseOutput):
-    """A node that performs a row filter by aggregating a given non-additive dimension.
-
-    This is designed to filter a dataset down to singular non-additive time dimension values by aggregating
-    the time dimension with MAX/MIN then joining back to the original dataset on that aggregated value.
-    Additionally, an optional sequence of entities can be passed in to group by and join on during the filtering.
-
-    For example, if we have a data set that includes "account_balances,user,date", we can build a
-    "latest_account_balance_by_user" data set using this node by filtering the data set such that for each user,
-    we get the MAX(date) and return the account_balance corresponding to that date.
-
-    Data transformation example,
-    | date       | account_balance | user |                             | date       | account_balance | user |
-    |:-----------|-----------------|-----:|     entity_specs:       |:-----------|-----------------|-----:|
-    | 2019-12-31 |            1000 |    u1|       - user                | 2020-01-03 |            2000 |    u1|
-    | 2020-01-03 |            2000 |    u1| ->  time_dimension_spec: -> | 2020-01-12              1500 |    u2|
-    | 2020-01-09 |            3000 |    u2|       - date                | 2020-01-12 |            1000 |    u3|
-    | 2020-01-12 |            1500 |    u2|     agg_by_function:
-    | 2020-01-12 |            1000 |    u3|       - MAX
-
-
-    Similarly, if we don't provide any entity_specs, it would end up performing the aggregation filter only on the
-    time_dimension_spec without any grouping by any entities.
-
-    Data transformation example,
-    | date       | account_balance | user |                             | date       | account_balance |
-    |:-----------|-----------------|-----:|     entity_specs:       |:-----------|----------------:|
-    | 2019-12-31 |            1000 |    u1|                             | 2020-01-12 |            2500 |
-    | 2020-01-03 |            2000 |    u1| ->  time_dimension_spec: ->
-    | 2020-01-09 |            3000 |    u2|       - date
-    | 2020-01-12 |            1500 |    u2|     agg_by_function:
-    | 2020-01-12 |            1000 |    u3|       - MAX
-
-
-    Additionally, we can aggregate against 'windows' of a dataset. For example, if we group by a non_additive time dimension,
-    we perform the MIN/MAX filtering on the granularity window provided in that group by instead of the full time range.
-
-    Data transformation example,
-    | date       | account_balance | user |                                  | date       | account_balance |
-    |:-----------|-----------------|-----:|     entity_specs:            |:-----------|----------------:|
-    | 2019-12-31 |            1500 |    u1|     time_dimension_spec:         | 2019-12-31 |            1500 |
-    | 2020-01-03 |            2000 |    u1| ->    - date                  -> | 2020-01-07 |            3000 |
-    | 2020-01-09 |            3000 |    u2|     agg_by_function:             | 2020-01-14 |            3250 |
-    | 2020-01-12 |            1500 |    u2|       - MIN
-    | 2020-01-14 |            1250 |    u3|     queried_time_dimension_spec:
-    | 2020-01-14 |            2000 |    u2|       - date__week
-    | 2020-01-15 |            4000 |    u1|
-    """
-
-    def __init__(
-        self,
-        parent_node: BaseOutput,
-        entity_specs: Sequence[LinklessEntitySpec],
-        time_dimension_spec: TimeDimensionSpec,
-        agg_by_function: AggregationType,
-        queried_time_dimension_spec: Optional[TimeDimensionSpec] = None,
-    ) -> None:
-        """Constructor.
-
-        Args:
-            parent_node: node with standard output
-            entity_specs: the entities to group the join by
-            time_dimension_spec: the time dimension used for row filtering via an aggregation
-            agg_by_function: the aggregation function used on the time dimension
-            queried_time_dimension_spec: The group by provided in the query used to build the windows we want to filter on.
-        """
-        self._parent_node = parent_node
-        self._entity_specs = tuple(entity_specs)
-        self._time_dimension_spec = time_dimension_spec
-        self._agg_by_function = agg_by_function
-        self._queried_time_dimension_spec = queried_time_dimension_spec
-
-        # Doing a list comprehension throws a type error, so doing it this way.
-        parent_nodes: Sequence[DataflowPlanNode] = [self._parent_node]
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=parent_nodes)
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_SEMI_ADDITIVE_JOIN_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_semi_additive_join_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return f"""Join on {self.agg_by_function.name}({self.time_dimension_spec.element_name}) and {[i.element_name for i in self.entity_specs]} grouping by {self.queried_time_dimension_spec.element_name if self.queried_time_dimension_spec else None}"""
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        return self._parent_node
-
-    @property
-    def entity_specs(self) -> Sequence[LinklessEntitySpec]:  # noqa: D
-        return self._entity_specs
-
-    @property
-    def time_dimension_spec(self) -> TimeDimensionSpec:  # noqa: D
-        return self._time_dimension_spec
-
-    @property
-    def agg_by_function(self) -> AggregationType:  # noqa: D
-        return self._agg_by_function
-
-    @property
-    def queried_time_dimension_spec(self) -> Optional[TimeDimensionSpec]:  # noqa: D
-        return self._queried_time_dimension_spec
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return super().displayed_properties
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        if not isinstance(other_node, self.__class__):
-            return False
-
-        return (
-            isinstance(other_node, self.__class__)
-            and other_node.entity_specs == self.entity_specs
-            and other_node.time_dimension_spec == self.time_dimension_spec
-            and other_node.agg_by_function == self.agg_by_function
-            and other_node.queried_time_dimension_spec == self.queried_time_dimension_spec
-        )
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> SemiAdditiveJoinNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-
-        return SemiAdditiveJoinNode(
-            parent_node=new_parent_nodes[0],
-            entity_specs=self.entity_specs,
-            time_dimension_spec=self.time_dimension_spec,
-            agg_by_function=self.agg_by_function,
-            queried_time_dimension_spec=self.queried_time_dimension_spec,
-        )
-
-
 class ComputedMetricsOutput(BaseOutput, ABC):
     """A node that outputs data that contains metrics computed from measures."""
 
     pass
-
-
-class JoinToTimeSpineNode(BaseOutput, ABC):
-    """Join parent dataset to time spine dataset."""
-
-    def __init__(
-        self,
-        parent_node: BaseOutput,
-        requested_agg_time_dimension_specs: Sequence[TimeDimensionSpec],
-        use_custom_agg_time_dimension: bool,
-        join_type: SqlJoinType,
-        time_range_constraint: Optional[TimeRangeConstraint] = None,
-        offset_window: Optional[MetricTimeWindow] = None,
-        offset_to_grain: Optional[TimeGranularity] = None,
-    ) -> None:
-        """Constructor.
-
-        Args:
-            parent_node: Node that returns desired dataset to join to time spine.
-            requested_agg_time_dimension_specs: Time dimensions requested in query.
-            use_custom_agg_time_dimension: Indicates if agg_time_dimension should be used in join. If false, uses metric_time.
-            time_range_constraint: Time range to constrain the time spine to.
-            offset_window: Time window to offset the parent dataset by when joining to time spine.
-            offset_to_grain: Granularity period to offset the parent dataset to when joining to time spine.
-
-        Passing both offset_window and offset_to_grain not allowed.
-        """
-        assert not (
-            offset_window and offset_to_grain
-        ), "Can't set both offset_window and offset_to_grain when joining to time spine. Choose one or the other."
-        assert (
-            len(requested_agg_time_dimension_specs) > 0
-        ), "Must have at least one value in requested_agg_time_dimension_specs for JoinToTimeSpineNode."
-
-        self._parent_node = parent_node
-        self._requested_agg_time_dimension_specs = tuple(requested_agg_time_dimension_specs)
-        self._use_custom_agg_time_dimension = use_custom_agg_time_dimension
-        self._offset_window = offset_window
-        self._offset_to_grain = offset_to_grain
-        self._time_range_constraint = time_range_constraint
-        self._join_type = join_type
-
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(self._parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_JOIN_TO_TIME_SPINE_ID_PREFIX
-
-    @property
-    def requested_agg_time_dimension_specs(self) -> Sequence[TimeDimensionSpec]:
-        """Time dimension specs to use when creating time spine table."""
-        return self._requested_agg_time_dimension_specs
-
-    @property
-    def use_custom_agg_time_dimension(self) -> bool:
-        """Whether or not metric_time was included in the query."""
-        return self._use_custom_agg_time_dimension
-
-    @property
-    def time_range_constraint(self) -> Optional[TimeRangeConstraint]:
-        """Time range constraint to apply when querying time spine table."""
-        return self._time_range_constraint
-
-    @property
-    def offset_window(self) -> Optional[MetricTimeWindow]:
-        """Time range constraint to apply when querying time spine table."""
-        return self._offset_window
-
-    @property
-    def offset_to_grain(self) -> Optional[TimeGranularity]:
-        """Time range constraint to apply when querying time spine table."""
-        return self._offset_to_grain
-
-    @property
-    def join_type(self) -> SqlJoinType:
-        """Join type to use when joining to time spine."""
-        return self._join_type
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_join_to_time_spine_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return """Join to Time Spine Dataset"""
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return tuple(super().displayed_properties) + (
-            DisplayedProperty("requested_agg_time_dimension_specs", self._requested_agg_time_dimension_specs),
-            DisplayedProperty("use_custom_agg_time_dimension", self._use_custom_agg_time_dimension),
-            DisplayedProperty("time_range_constraint", self._time_range_constraint),
-            DisplayedProperty("offset_window", self._offset_window),
-            DisplayedProperty("offset_to_grain", self._offset_to_grain),
-            DisplayedProperty("join_type", self._join_type),
-        )
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        return self._parent_node
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return (
-            isinstance(other_node, self.__class__)
-            and other_node.time_range_constraint == self.time_range_constraint
-            and other_node.offset_window == self.offset_window
-            and other_node.offset_to_grain == self.offset_to_grain
-            and other_node.requested_agg_time_dimension_specs == self.requested_agg_time_dimension_specs
-            and other_node.use_custom_agg_time_dimension == self.use_custom_agg_time_dimension
-            and other_node.join_type == self.join_type
-        )
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> JoinToTimeSpineNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return JoinToTimeSpineNode(
-            parent_node=new_parent_nodes[0],
-            requested_agg_time_dimension_specs=self.requested_agg_time_dimension_specs,
-            use_custom_agg_time_dimension=self.use_custom_agg_time_dimension,
-            time_range_constraint=self.time_range_constraint,
-            offset_window=self.offset_window,
-            offset_to_grain=self.offset_to_grain,
-            join_type=self.join_type,
-        )
-
-
-class ComputeMetricsNode(ComputedMetricsOutput):
-    """A node that computes metrics from input measures. Dimensions / entities are passed through."""
-
-    def __init__(self, parent_node: BaseOutput, metric_specs: Sequence[MetricSpec]) -> None:  # noqa: D
-        """Constructor.
-
-        Args:
-            parent_node: Node where data is coming from.
-            metric_specs: The specs for the metrics that this should compute.
-        """
-        self._parent_node = parent_node
-        self._metric_specs = tuple(metric_specs)
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(self._parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_COMPUTE_METRICS_ID_PREFIX
-
-    @property
-    def metric_specs(self) -> Sequence[MetricSpec]:  # noqa: D
-        """The metric instances that this node is supposed to compute and should have in the output."""
-        return self._metric_specs
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_compute_metrics_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return """Compute Metrics via Expressions"""
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return tuple(super().displayed_properties) + tuple(
-            DisplayedProperty("metric_spec", metric_spec) for metric_spec in self._metric_specs
-        )
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        return self._parent_node
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        if not isinstance(other_node, self.__class__):
-            return False
-
-        if other_node.metric_specs != self.metric_specs:
-            return False
-
-        return isinstance(other_node, self.__class__) and other_node.metric_specs == self.metric_specs
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> ComputeMetricsNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return ComputeMetricsNode(
-            parent_node=new_parent_nodes[0],
-            metric_specs=self.metric_specs,
-        )
-
-
-class OrderByLimitNode(ComputedMetricsOutput):
-    """A node that re-orders the input data with a limit."""
-
-    def __init__(
-        self,
-        order_by_specs: Sequence[OrderBySpec],
-        parent_node: Union[BaseOutput, ComputedMetricsOutput],
-        limit: Optional[int] = None,
-    ) -> None:
-        """Constructor.
-
-        Args:
-            order_by_specs: describes how to order the incoming data.
-            limit: number of rows to limit.
-            parent_node: self-explanatory.
-        """
-        self._order_by_specs = tuple(order_by_specs)
-        self._limit = limit
-        self._parent_node = parent_node
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(self._parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_ORDER_BY_LIMIT_ID_PREFIX
-
-    @property
-    def order_by_specs(self) -> Sequence[OrderBySpec]:
-        """The elements that this node should order the input data."""
-        return self._order_by_specs
-
-    @property
-    def limit(self) -> Optional[int]:
-        """The number of rows to limit by."""
-        return self._limit
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_order_by_limit_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return f"Order By {[order_by_spec.instance_spec.qualified_name for order_by_spec in self._order_by_specs]}" + (
-            f" Limit {self._limit}" if self.limit else ""
-        )
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return (
-            tuple(super().displayed_properties)
-            + tuple(DisplayedProperty("order_by_spec", order_by_spec) for order_by_spec in self._order_by_specs)
-            + (DisplayedProperty("limit", str(self.limit)),)
-        )
-
-    @property
-    def parent_node(self) -> Union[BaseOutput, ComputedMetricsOutput]:  # noqa: D
-        return self._parent_node
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return (
-            isinstance(other_node, self.__class__)
-            and other_node.order_by_specs == self.order_by_specs
-            and other_node.limit == self.limit
-        )
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> OrderByLimitNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-
-        return OrderByLimitNode(
-            parent_node=new_parent_nodes[0],
-            order_by_specs=self.order_by_specs,
-            limit=self.limit,
-        )
-
-
-class MetricTimeDimensionTransformNode(BaseOutput):
-    """A node transforms the input data set so that it contains the metric time dimension and relevant measures.
-
-    The metric time dimension is used later to aggregate all measures in the data set.
-
-    Input: a data set containing measures along with the associated aggregation time dimension.
-
-    Output: a data set similar to the input data set, but includes the configured aggregation time dimension as the
-    metric time dimension and only contains measures that are defined to use it.
-    """
-
-    def __init__(  # noqa: D
-        self,
-        parent_node: BaseOutput,
-        aggregation_time_dimension_reference: TimeDimensionReference,
-    ) -> None:
-        self._aggregation_time_dimension_reference = aggregation_time_dimension_reference
-        self._parent_node = parent_node
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_SET_MEASURE_AGGREGATION_TIME
-
-    @property
-    def aggregation_time_dimension_reference(self) -> TimeDimensionReference:
-        """The time dimension that measures in the input should be aggregated to."""
-        return self._aggregation_time_dimension_reference
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_metric_time_dimension_transform_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return f"Metric Time Dimension '{self.aggregation_time_dimension_reference.element_name}'" ""
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return tuple(super().displayed_properties) + (
-            DisplayedProperty("aggregation_time_dimension", self.aggregation_time_dimension_reference.element_name),
-        )
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        return self._parent_node
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return (
-            isinstance(other_node, self.__class__)
-            and other_node.aggregation_time_dimension_reference == self.aggregation_time_dimension_reference
-        )
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> MetricTimeDimensionTransformNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return MetricTimeDimensionTransformNode(
-            parent_node=new_parent_nodes[0],
-            aggregation_time_dimension_reference=self.aggregation_time_dimension_reference,
-        )
 
 
 class SinkNodeVisitor(Generic[VisitorOutputT], ABC):
@@ -981,493 +216,6 @@ class SinkOutput(DataflowPlanNode, ABC):
     @abstractmethod
     def parent_node(self) -> BaseOutput:  # noqa: D
         pass
-
-
-class WriteToResultDataframeNode(SinkOutput):
-    """A node where incoming data gets written to a dataframe."""
-
-    def __init__(self, parent_node: BaseOutput) -> None:  # noqa: D
-        self._parent_node = parent_node
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_WRITE_TO_RESULT_DATAFRAME_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_write_to_result_dataframe_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return """Write to Dataframe"""
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        assert len(self.parent_nodes) == 1
-        return self._parent_node
-
-    def accept_sink_node_visitor(self, visitor: SinkNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_write_to_result_dataframe_node(self)
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return isinstance(other_node, self.__class__)
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> WriteToResultDataframeNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return WriteToResultDataframeNode(parent_node=new_parent_nodes[0])
-
-
-class WriteToResultTableNode(SinkOutput):
-    """A node where incoming data gets written to a table."""
-
-    def __init__(  # noqa: D
-        self,
-        parent_node: BaseOutput,
-        output_sql_table: SqlTable,
-    ) -> None:
-        """Constructor.
-
-        Args:
-            parent_node: node that outputs the computed metrics.
-            output_sql_table: the table where the computed metrics should be written to.
-        """
-        self._parent_node = parent_node
-        self._output_sql_table = output_sql_table
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_WRITE_TO_RESULT_DATAFRAME_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_write_to_result_table_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return """Write to Table"""
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        assert len(self.parent_nodes) == 1
-        return self._parent_node
-
-    def accept_sink_node_visitor(self, visitor: SinkNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_write_to_result_table_node(self)
-
-    @property
-    def output_sql_table(self) -> SqlTable:  # noqa: D
-        return self._output_sql_table
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return isinstance(other_node, self.__class__) and other_node.output_sql_table == self.output_sql_table
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> WriteToResultTableNode:  # noqa: D
-        return WriteToResultTableNode(
-            parent_node=new_parent_nodes[0],
-            output_sql_table=self.output_sql_table,
-        )
-
-
-class FilterElementsNode(BaseOutput):
-    """Only passes the listed elements."""
-
-    def __init__(  # noqa: D
-        self,
-        parent_node: BaseOutput,
-        include_specs: InstanceSpecSet,
-        replace_description: Optional[str] = None,
-        distinct: bool = False,
-    ) -> None:
-        self._include_specs = include_specs
-        self._replace_description = replace_description
-        self._parent_node = parent_node
-        self._distinct = distinct
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_PASS_FILTER_ELEMENTS_ID_PREFIX
-
-    @property
-    def include_specs(self) -> InstanceSpecSet:
-        """Returns the specs for the elements that it should pass."""
-        return self._include_specs
-
-    @property
-    def distinct(self) -> bool:
-        """True if you only want the distinct values for the selected specs."""
-        return self._distinct
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_pass_elements_filter_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        if self._replace_description:
-            return self._replace_description
-
-        return f"Pass Only Elements: {mf_pformat([x.qualified_name for x in self._include_specs.all_specs])}"
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        additional_properties: Tuple[DisplayedProperty, ...] = ()
-        if not self._replace_description:
-            additional_properties = tuple(
-                DisplayedProperty("include_spec", include_spec) for include_spec in self._include_specs.all_specs
-            ) + (
-                DisplayedProperty("distinct", self._distinct),
-            )
-        return tuple(super().displayed_properties) + additional_properties
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        return self._parent_node
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return (
-            isinstance(other_node, self.__class__)
-            and other_node.include_specs == self.include_specs
-            and other_node.distinct == self.distinct
-        )
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> FilterElementsNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return FilterElementsNode(
-            parent_node=new_parent_nodes[0],
-            include_specs=self.include_specs,
-            distinct=self.distinct,
-            replace_description=self._replace_description,
-        )
-
-
-class WhereConstraintNode(AggregatedMeasuresOutput):
-    """Remove rows using a WHERE clause."""
-
-    def __init__(  # noqa: D
-        self,
-        parent_node: BaseOutput,
-        where_constraint: WhereFilterSpec,
-    ) -> None:
-        self._where = where_constraint
-        self.parent_node = parent_node
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_WHERE_CONSTRAINT_ID_PREFIX
-
-    @property
-    def where(self) -> WhereFilterSpec:
-        """Returns the specs for the elements that it should pass."""
-        return self._where
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_where_constraint_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        # Can't put the where condition here as it can cause rendering issues when there are SQL execution parameters.
-        # e.g. "Constrain Output with WHERE listing__country = :1"
-        return "Constrain Output with WHERE"
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return tuple(super().displayed_properties) + (DisplayedProperty("where_condition", self.where),)
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return isinstance(other_node, self.__class__) and other_node.where == self.where
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> WhereConstraintNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return WhereConstraintNode(
-            parent_node=new_parent_nodes[0],
-            where_constraint=self.where,
-        )
-
-
-class CombineAggregatedOutputsNode(ComputedMetricsOutput):
-    """Combines metrics from different nodes into a single output."""
-
-    def __init__(  # noqa: D
-        self,
-        parent_nodes: Sequence[Union[BaseOutput, ComputedMetricsOutput]],
-    ) -> None:
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=parent_nodes)
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_COMBINE_AGGREGATED_OUTPUTS_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_combine_aggregated_outputs_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return "Combine Aggregated Outputs"
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return isinstance(other_node, self.__class__)
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> CombineAggregatedOutputsNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return CombineAggregatedOutputsNode(parent_nodes=new_parent_nodes)
-
-
-class ConstrainTimeRangeNode(AggregatedMeasuresOutput, BaseOutput):
-    """Constrains the time range of the input data set.
-
-    For example, if the input data set had "sales by date", then this would restrict the data set so that it only
-    includes sales for a specific range of dates.
-    """
-
-    def __init__(  # noqa: D
-        self,
-        parent_node: BaseOutput,
-        time_range_constraint: TimeRangeConstraint,
-    ) -> None:
-        self._time_range_constraint = time_range_constraint
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_CONSTRAIN_TIME_RANGE_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_constrain_time_range_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return (
-            f"Constrain Time Range to [{self.time_range_constraint.start_time.isoformat()}, "
-            f"{self.time_range_constraint.end_time.isoformat()}]"
-        )
-
-    @property
-    def time_range_constraint(self) -> TimeRangeConstraint:  # noqa: D
-        return self._time_range_constraint
-
-    @property
-    def parent_node(self) -> DataflowPlanNode:  # noqa: D
-        assert len(self.parent_nodes) == 1
-        return self.parent_nodes[0]
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return tuple(super().displayed_properties) + (
-            DisplayedProperty("time_range_start", self.time_range_constraint.start_time.isoformat()),
-            DisplayedProperty("time_range_end", self.time_range_constraint.end_time.isoformat()),
-        )
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return isinstance(other_node, self.__class__) and self.time_range_constraint == other_node.time_range_constraint
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> ConstrainTimeRangeNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return ConstrainTimeRangeNode(
-            parent_node=new_parent_nodes[0],
-            time_range_constraint=self.time_range_constraint,
-        )
-
-
-class MinMaxNode(BaseOutput):
-    """Calculate the min and max of a single instance data set."""
-
-    def __init__(self, parent_node: BaseOutput) -> None:  # noqa: D
-        self._parent_node = parent_node
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_MIN_MAX_ID_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_min_max_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return "Calculate min and max"
-
-    @property
-    def parent_node(self) -> BaseOutput:  # noqa: D
-        return self._parent_node
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return isinstance(other_node, self.__class__)
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> MinMaxNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return MinMaxNode(parent_node=new_parent_nodes[0])
-
-
-class AddGeneratedUuidColumnNode(BaseOutput):
-    """Adds a UUID column."""
-
-    def __init__(self, parent_node: BaseOutput) -> None:  # noqa: D
-        super().__init__(node_id=self.create_unique_id(), parent_nodes=(parent_node,))
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_ADD_UUID_COLUMN_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_add_generated_uuid_column_node(self)
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return "Adds an internally generated UUID column"
-
-    @property
-    def parent_node(self) -> DataflowPlanNode:  # noqa: D
-        assert len(self.parent_nodes) == 1
-        return self.parent_nodes[0]
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return super().displayed_properties
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return isinstance(other_node, self.__class__)
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> AddGeneratedUuidColumnNode:  # noqa: D
-        assert len(new_parent_nodes) == 1
-        return AddGeneratedUuidColumnNode(parent_node=new_parent_nodes[0])
-
-
-class JoinConversionEventsNode(BaseOutput):
-    """Builds a data set containing successful conversion events."""
-
-    def __init__(
-        self,
-        base_node: BaseOutput,
-        base_time_dimension_spec: TimeDimensionSpec,
-        conversion_node: BaseOutput,
-        conversion_measure_spec: MeasureSpec,
-        conversion_time_dimension_spec: TimeDimensionSpec,
-        unique_identifier_keys: Sequence[InstanceSpec],
-        entity_spec: EntitySpec,
-        window: Optional[MetricTimeWindow] = None,
-        constant_properties: Optional[Sequence[ConstantPropertySpec]] = None,
-    ) -> None:
-        """Constructor.
-
-        Args:
-            base_node: node containing dataset for computing base events.
-            base_time_dimension_spec: time dimension for the base events to compute against.
-            conversion_node: node containing dataset to join base node for computing conversion events.
-            conversion_measure_spec: expose this measure in the resulting dataset for aggregation.
-            conversion_time_dimension_spec: time dimension for the conversion events to compute against.
-            unique_identifier_keys: columns to uniquely identify each conversion event.
-            entity_spec: the specific entity in which the conversion is happening for.
-            window: time range bound for when a conversion is still considered valid (default: INF).
-            constant_properties: optional set of elements (either dimension/entity) to join the base
-                                 event to the conversion event.
-        """
-        self._base_node = base_node
-        self._conversion_node = conversion_node
-        self._base_time_dimension_spec = base_time_dimension_spec
-        self._conversion_measure_spec = conversion_measure_spec
-        self._conversion_time_dimension_spec = conversion_time_dimension_spec
-        self._unique_identifier_keys = tuple(unique_identifier_keys)
-        self._entity_spec = entity_spec
-        self._window = window
-        self._constant_properties = constant_properties
-        super().__init__(
-            node_id=self.create_unique_id(),
-            parent_nodes=(
-                base_node,
-                conversion_node,
-            ),
-        )
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D
-        return StaticIdPrefix.DATAFLOW_NODE_JOIN_CONVERSION_EVENTS_PREFIX
-
-    def accept(self, visitor: DataflowPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D
-        return visitor.visit_join_conversion_events_node(self)
-
-    @property
-    def base_node(self) -> DataflowPlanNode:  # noqa: D
-        return self._base_node
-
-    @property
-    def conversion_node(self) -> DataflowPlanNode:  # noqa: D
-        return self._conversion_node
-
-    @property
-    def conversion_measure_spec(self) -> MeasureSpec:  # noqa: D
-        return self._conversion_measure_spec
-
-    @property
-    def base_time_dimension_spec(self) -> TimeDimensionSpec:  # noqa: D
-        return self._base_time_dimension_spec
-
-    @property
-    def conversion_time_dimension_spec(self) -> TimeDimensionSpec:  # noqa: D
-        return self._conversion_time_dimension_spec
-
-    @property
-    def unique_identifier_keys(self) -> Sequence[InstanceSpec]:  # noqa: D
-        return self._unique_identifier_keys
-
-    @property
-    def entity_spec(self) -> EntitySpec:  # noqa: D
-        return self._entity_spec
-
-    @property
-    def window(self) -> Optional[MetricTimeWindow]:  # noqa: D
-        return self._window
-
-    @property
-    def constant_properties(self) -> Optional[Sequence[ConstantPropertySpec]]:  # noqa: D
-        return self._constant_properties
-
-    @property
-    def description(self) -> str:  # noqa: D
-        return f"Find conversions for {self.entity_spec.qualified_name} within the range of {f'{self.window.count} {self.window.granularity.value}' if self.window else 'INF'}"
-
-    @property
-    def displayed_properties(self) -> Sequence[DisplayedProperty]:  # noqa: D
-        return (
-            tuple(super().displayed_properties)
-            + (
-                DisplayedProperty("base_time_dimension_spec", self.base_time_dimension_spec),
-                DisplayedProperty("conversion_time_dimension_spec", self.conversion_time_dimension_spec),
-                DisplayedProperty("entity_spec", self.entity_spec),
-                DisplayedProperty("window", self.window),
-            )
-            + tuple(DisplayedProperty("unique_key_specs", unique_spec) for unique_spec in self.unique_identifier_keys)
-            + tuple(
-                DisplayedProperty("constant_property", constant_property)
-                for constant_property in self.constant_properties or []
-            )
-        )
-
-    def functionally_identical(self, other_node: DataflowPlanNode) -> bool:  # noqa: D
-        return (
-            isinstance(other_node, self.__class__)
-            and other_node.base_time_dimension_spec == self.base_time_dimension_spec
-            and other_node.conversion_time_dimension_spec == self.conversion_time_dimension_spec
-            and other_node.conversion_measure_spec == self.conversion_measure_spec
-            and other_node.unique_identifier_keys == self.unique_identifier_keys
-            and other_node.entity_spec == self.entity_spec
-            and other_node.window == self.window
-            and other_node.constant_properties == self.constant_properties
-        )
-
-    def with_new_parents(self, new_parent_nodes: Sequence[BaseOutput]) -> JoinConversionEventsNode:  # noqa: D
-        assert len(new_parent_nodes) == 2
-        return JoinConversionEventsNode(
-            base_node=new_parent_nodes[0],
-            base_time_dimension_spec=self.base_time_dimension_spec,
-            conversion_node=new_parent_nodes[1],
-            conversion_measure_spec=self.conversion_measure_spec,
-            conversion_time_dimension_spec=self.conversion_time_dimension_spec,
-            unique_identifier_keys=self.unique_identifier_keys,
-            entity_spec=self.entity_spec,
-            window=self.window,
-            constant_properties=self.constant_properties,
-        )
 
 
 class DataflowPlan(MetricFlowDag[SinkOutput]):
