+++ metricflow/plan_conversion/sql_join_builder.py
@@ -7,12 +7,10 @@ from dbt_semantic_interfaces.protocols.metric import MetricTimeWindow
 from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 from metricflow.assert_one_arg import assert_exactly_one_arg_set
-from metricflow.dataflow.dataflow_plan import (
-    JoinConversionEventsNode,
-    JoinDescription,
-    JoinOverTimeRangeNode,
-    JoinToTimeSpineNode,
-)
+from metricflow.dataflow.nodes.join_conversion_events import JoinConversionEventsNode
+from metricflow.dataflow.nodes.join_over_time import JoinOverTimeRangeNode
+from metricflow.dataflow.nodes.join_to_base import JoinDescription
+from metricflow.dataflow.nodes.join_to_time_spine import JoinToTimeSpineNode
 from metricflow.dataset.sql_dataset import SqlDataSet
 from metricflow.plan_conversion.sql_expression_builders import make_coalesced_expr
 from metricflow.sql.sql_exprs import (
