+++ metricflow/test/dataflow/optimizer/source_scan/test_source_scan_optimizer.py
@@ -9,28 +9,28 @@ from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 from metricflow.dataflow.builder.dataflow_plan_builder import DataflowPlanBuilder
 from metricflow.dataflow.dataflow_plan import (
-    AddGeneratedUuidColumnNode,
-    AggregateMeasuresNode,
-    CombineAggregatedOutputsNode,
-    ComputeMetricsNode,
-    ConstrainTimeRangeNode,
     DataflowPlan,
     DataflowPlanNode,
     DataflowPlanNodeVisitor,
-    FilterElementsNode,
-    JoinConversionEventsNode,
-    JoinOverTimeRangeNode,
-    JoinToBaseOutputNode,
-    JoinToTimeSpineNode,
-    MetricTimeDimensionTransformNode,
-    MinMaxNode,
-    OrderByLimitNode,
-    ReadSqlSourceNode,
-    SemiAdditiveJoinNode,
-    WhereConstraintNode,
-    WriteToResultDataframeNode,
-    WriteToResultTableNode,
 )
+from metricflow.dataflow.nodes.add_generated_uuid import AddGeneratedUuidColumnNode
+from metricflow.dataflow.nodes.aggregate_measures import AggregateMeasuresNode
+from metricflow.dataflow.nodes.combine_aggregated_outputs import CombineAggregatedOutputsNode
+from metricflow.dataflow.nodes.compute_metrics import ComputeMetricsNode
+from metricflow.dataflow.nodes.constrain_time import ConstrainTimeRangeNode
+from metricflow.dataflow.nodes.filter_elements import FilterElementsNode
+from metricflow.dataflow.nodes.join_conversion_events import JoinConversionEventsNode
+from metricflow.dataflow.nodes.join_over_time import JoinOverTimeRangeNode
+from metricflow.dataflow.nodes.join_to_base import JoinToBaseOutputNode
+from metricflow.dataflow.nodes.join_to_time_spine import JoinToTimeSpineNode
+from metricflow.dataflow.nodes.metric_time_transform import MetricTimeDimensionTransformNode
+from metricflow.dataflow.nodes.min_max import MinMaxNode
+from metricflow.dataflow.nodes.order_by_limit import OrderByLimitNode
+from metricflow.dataflow.nodes.read_sql_source import ReadSqlSourceNode
+from metricflow.dataflow.nodes.semi_additive_join import SemiAdditiveJoinNode
+from metricflow.dataflow.nodes.where_filter import WhereConstraintNode
+from metricflow.dataflow.nodes.write_to_dataframe import WriteToResultDataframeNode
+from metricflow.dataflow.nodes.write_to_table import WriteToResultTableNode
 from metricflow.dataflow.optimizer.source_scan.source_scan_optimizer import SourceScanOptimizer
 from metricflow.dataset.dataset import DataSet
 from metricflow.query.query_parser import MetricFlowQueryParser
