+++ metricflow/test/plan_conversion/test_dataflow_to_sql_plan.py
@@ -13,22 +13,21 @@ from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 from metricflow.dag.mf_dag import DagId
 from metricflow.dataflow.builder.dataflow_plan_builder import DataflowPlanBuilder
 from metricflow.dataflow.dataflow_plan import (
-    AggregateMeasuresNode,
     BaseOutput,
-    CombineAggregatedOutputsNode,
-    ComputeMetricsNode,
-    ConstrainTimeRangeNode,
     DataflowPlan,
-    FilterElementsNode,
-    JoinDescription,
-    JoinToBaseOutputNode,
-    JoinToTimeSpineNode,
-    MetricTimeDimensionTransformNode,
-    OrderByLimitNode,
-    SemiAdditiveJoinNode,
-    WhereConstraintNode,
-    WriteToResultDataframeNode,
 )
+from metricflow.dataflow.nodes.aggregate_measures import AggregateMeasuresNode
+from metricflow.dataflow.nodes.combine_aggregated_outputs import CombineAggregatedOutputsNode
+from metricflow.dataflow.nodes.compute_metrics import ComputeMetricsNode
+from metricflow.dataflow.nodes.constrain_time import ConstrainTimeRangeNode
+from metricflow.dataflow.nodes.filter_elements import FilterElementsNode
+from metricflow.dataflow.nodes.join_to_base import JoinDescription, JoinToBaseOutputNode
+from metricflow.dataflow.nodes.join_to_time_spine import JoinToTimeSpineNode
+from metricflow.dataflow.nodes.metric_time_transform import MetricTimeDimensionTransformNode
+from metricflow.dataflow.nodes.order_by_limit import OrderByLimitNode
+from metricflow.dataflow.nodes.semi_additive_join import SemiAdditiveJoinNode
+from metricflow.dataflow.nodes.where_filter import WhereConstraintNode
+from metricflow.dataflow.nodes.write_to_dataframe import WriteToResultDataframeNode
 from metricflow.filters.time_constraint import TimeRangeConstraint
 from metricflow.plan_conversion.dataflow_to_sql import DataflowToSqlQueryPlanConverter
 from metricflow.protocols.sql_client import SqlClient
